Simplifies test package configuration
For a while, the default package where AS creates test files is of the form com.example.android, whereas the instrumentation package is com.example.android.tests. Fork has been defaulting to the instrumentation package to search for tests, however this is not ideal as it will be not working with default projects. From now on, Fork will just scan the simple form (com.example.android) and most projects won't even need to specify a test package, since there's an alignment between app package (Fork already knows that) and tests package.