Revert "Ensures test crashed before rescheduling"
This commit was rescheduling a failed run at the point of failure. As part of the rescheduling is the deletion of the test report. Moving the scheduling at the point of failure, making this impossible to delete the generated report, which gets created in the very end of the run.