Restores equals and hashCode in TestCaseEvent to match only testMetod and testClass, ignoring any other field in the class.
When rescheduling tests (via the RetryListener) and when capturing test properties (via ForkXmlTestRunListener#getPropertiesAttributes() )  a new TestCaseEvent is synthetized from the ddmlib's TestIdentifier: this is a pojo class with only test class and method, so any further field in TestCaseEvent is initialized with defaults.
PoolTestCaseFailureAccumulator was not finding the correct tests in the pools as the .equal() in TestCaseEvent was comparing a full object with a synthetized one laking of some fields, returning false all the times

Note: In Fork this affects test that have custom annotations being converted to test properties.