Uses TestIdentifier when only test class name and method are necessary
This solves a bug where in RetryListener we were synthetizing a TestCaseEvent (that could contains test metadata) from a TestIdentifier (pojo class with only test class and method name).
In the test case accumulator we were then comparing a fully generated TestCaseEvent to the synthetized counterpart, and the equals was always false in presence of extra test properties.
Note: this was happening only if a test was annotated with custom properties via the TestProperties annotation.